turn about component into a page
