--- client/modules/IDE/components/About.stories.jsx
@@ -1,8 +1,0 @@
-import About from './About';
-
-export default {
-  title: 'IDE/About',
-  component: About
-};
-
-export const Default = {};
